Updated hoc`s Changed file names. Refactored AuthProvider.hoc
