+++ frontend/src/hook/useAuth.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthContext } from "../hoc/AuthProvider";
+import { AuthContext } from "../hoc/AuthProvider.hoc";
 
 export function useAuth() {
   return useContext(AuthContext);
